requirements.txt
# Documentation Toolkit - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Document processing
python-docx>=1.0.0
python-pptx>=0.6.21
pypdf>=3.0.0

# Image processing and OCR
Pillow>=10.0.0
pytesseract>=0.3.10
